+++ Cloud/terraform/modules/aws/variables.tf
@@ -1,0 +1,5 @@
+variable "bucket_name_prefix" {
+  description = "Prefix for the AWS S3 bucket name"
+  type        = string
+  default     = "robot-cloud-aws"
+}