+++ Cloud/terraform/modules/gcp/main.tf
@@ -1,8 +1,9 @@
 # GCP resources configuration
 
 resource "google_storage_bucket" "example" {
-  name     = "example-bucket-${random_id.bucket_id.hex}"
+  name     = "${var.bucket_name_prefix}-${random_id.bucket_id.hex}"
   location = var.gcp_region
+  project  = var.gcp_project
 }
 
 resource "random_id" "bucket_id" {