+++ Cloud/terraform/modules/gcp/outputs.tf
@@ -1,0 +1,4 @@
+output "bucket_name" {
+  description = "GCP storage bucket name"
+  value       = google_storage_bucket.example.name
+}